+++ src/__tests__/helpers.test.js
@@ -1,24 +1,24 @@
-const validateURI = require("../index.js");
+const tidyURI = require("../helpers.js");
 
-test("validateURI() can tidy URI", () => {
+test("tidyURI() can tidy URI", () => {
   const site = "sodiumhalogen.com";
   const siteHttp = "http://sodiumhalogen.com";
   const siteHttp2 = "htp://sodiumhalogen.com";
   const siteHttp3 = "htt://sodiumhalogen.com";
   const siteHttps = "https://sodiumhalogen.com";
 
-  const updatedSite = validateURI(site);
+  const updatedSite = tidyURI(site);
   expect(updatedSite).toBe("http://sodiumhalogen.com");
 
-  const updatedSiteHttp = validateURI(siteHttp);
+  const updatedSiteHttp = tidyURI(siteHttp);
   expect(updatedSiteHttp).toBe("http://sodiumhalogen.com");
 
-  const updatedSite2 = validateURI(siteHttp2);
+  const updatedSite2 = tidyURI(siteHttp2);
   expect(updatedSite2).toBe("http://sodiumhalogen.com");
 
-  const updatedSite3 = validateURI(siteHttp3);
+  const updatedSite3 = tidyURI(siteHttp3);
   expect(updatedSite2).toBe("http://sodiumhalogen.com");
 
-  const updatedSiteHttps = validateURI(siteHttps);
+  const updatedSiteHttps = tidyURI(siteHttps);
   expect(updatedSiteHttps).toBe("https://sodiumhalogen.com");
 });
